+++ code/tests/F0.CodeAnalysis.CSharp.SourceGenerator.Benchmarking.Tests/Benchmarking/TestCSharpSourceGenerator.cs
@@ -91,9 +91,7 @@ internal sealed class TestCSharpSourceGenerator : ISourceGenerator
 			context.AddSource(hintName, sourceText.ToString());
 		}
 
-		ImmutableDictionary<string, string?>.Builder builder = ImmutableDictionary.CreateBuilder<string, string?>();
-		builder.Add("Name", "Value");
-		ImmutableDictionary<string, string?> properties = builder.ToImmutable();
+		ImmutableDictionary<string, string?> properties = CreateProperties();
 
 		foreach (SyntaxNode node in receiver.Nodes)
 		{
@@ -104,5 +102,16 @@ internal sealed class TestCSharpSourceGenerator : ISourceGenerator
 				context.ReportDiagnostic(diagnostic);
 			}
 		}
+	}
+
+	private static ImmutableDictionary<string, string?> CreateProperties()
+	{
+		ImmutableDictionary<string, string?>.Builder builder = ImmutableDictionary.CreateBuilder<string, string?>();
+		builder.Add("Zero", "One");
+		builder.Add("Two", "Three");
+		builder.Add("Four", "Five");
+		builder.Add("Six", "Seven");
+		builder.Add("Eight", "Nine");
+		return builder.ToImmutable();
 	}
 }