+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Added _C#_ benchmarking types for `DiagnosticAnalyzer`.
 
 ## v0.2.0 (2022-07-31)
 - Added _C#_ benchmarking types for `ISourceGenerator`.