+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## vNext
 - Added _C#_ benchmarking types for `DiagnosticAnalyzer`.
+- Fixed inspection order of `Diagnostic.Properties`.
 
 ## v0.2.0 (2022-07-31)
 - Added _C#_ benchmarking types for `ISourceGenerator`.